// src/app/models/Session.ts
export class Session {
    id: number = 0;
    startedAt: string = "";
    userId: number = 0;
    userNr: number = 0;
    username: string = "";
    name: string = "";
    surname: string = "";
    userIsAdmin: number = 0;
    userERPToken: string = "";
    userLogoId: number = 0;
    worksapceId: number = 0;
    workspaceTitle: string = "";
    workspaceNr: number = 0;
    workspaceFirmNr: number = 0;
    workspacePerNr: number = 0;
    deviceId: number = 0;
    deviceNr: number = 0;
    deviceIdentifier: string = "";
    deviceDefinition: string = "";
}